presentation-api/controlling-ua/PresentationRequest_success.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Presentation API PresentationRequest for Controlling User Agent (Success)</title>
<link rel="author" title="Franck William Taffo" href="http://www.fokus.fraunhofer.de">
<link rel="help" href="http://w3c.github.io/presentation-api/#dfn-controlling-user-agent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>

    var client_id = String(new Date().getTime()) + String(Math.floor(Math.random() * 1e5));
    var presentation_url = 'https://google.com/cast#__castAppId__=CC1AD845/__castClientId__=' + client_id;

    /**
     *  Test if PresentationRequest constructor returns a SyntaxError() by missing presentation URL
     */

    test(function() {
        try{
            var request = new PresentationRequest(presentation_url);
            assert_true(request instanceof PresentationRequest);
        }
        catch (ex){
            assert_unreached("Call PresentationRequest() constructor with valid presentation URL....????");
        }
    }, 'Call PresentationRequest() constructor with valid presentation URL. No Exception expected.');

</script>
